add support for nginx/unicorn
modified puppet::master class for using $webserver.
this can be httpd for apache/passanger or nginx for nginx/unicorn.
nginx is currently only tested on centos7, but should work on evers OS with systemd

the nginx vhost definition is currently a simple file ressource, this should be changed
to nginx::resource:vhost in the future